+++ UDV-Core/src/Extensions/Contribute/src/ContributeController.js
@@ -9,6 +9,7 @@ import { CreateDocument }  from './CreateDocument.js';
 import { UpdateDocument }   from './UpdateDocument.js';
 import "./Contribute.css";
 import { MAIN_LOOP_EVENTS } from 'itowns';
+import { removeEmptyValues } from '../../../Utils/DataProcessing/DataProcessing';
 
 /**
  *
@@ -220,6 +221,7 @@ export function ContributeController(documentController, requestService){
   this.documentUpdate = function documentUpdate(){
 
     this.updatedData = new FormData(document.getElementById(this.documentUpdate.updateFormId));
+    this.updatedData = removeEmptyValues(this.updatedData);
 
     //get current doc data and id
     var currentDoc = this.documentController.getCurrentDoc();
