+++ UDV-Core/src/Utils/DataProcessing/DataProcessing.js
@@ -1,0 +1,12 @@
+export function removeEmptyValues(formData) {
+  let emptyKeys = [];
+  formData.forEach((value, key) => {
+    if (!value) {
+      emptyKeys.push(key)
+    }
+  });
+  emptyKeys.forEach((key) => {
+    formData.delete(key);
+  });
+  return formData;
+}
